test: split App.coverage.test.ts into focused files under 200 LOC each
Split the 898-line App.coverage.test.ts into 8 focused files:
- App.coverage.test-utils.ts — shared MockEventSource and baseState
- App.coverage.helpers.test.ts — ANSI/string/date helpers (109 LOC)
- App.coverage.parsing.test.ts — parsing/health/artifact helpers (181 LOC)
- App.coverage.comparison.test.ts — ArtifactComparisonDialog (89 LOC)
- App.coverage.integration-qa.test.ts — QA badge tests (149 LOC)
- App.coverage.integration-app.test.ts — steer/stop/resume/panel (154 LOC)
- App.coverage.integration-sidebar.test.ts — sidebar/session tests (167 LOC)
- App.coverage.panels.test.ts — ActivityPanel/DocContent/HealthPanel (58 LOC)

Aloop-Agent: unknown
Aloop-Iteration: 1
Aloop-Session: orchestrator-20260321-172932-issue-189-20260324-080746
